Refactor a bit.
Move all logic out of cmd/root.go into a package.
